Add :core:database, change type of event id to Long

+++ feature/events/src/main/java/com/arshapshap/events/domain/EventsInteractor.kt
@@ -12,11 +12,11 @@ class EventsInteractor @Inject constructor(
         return repository.getEvents()
     }
 
-    internal suspend fun getEventById(id: Int): Event? {
+    internal suspend fun getEventById(id: Long): Event? {
         return repository.getEventById(id)
     }
 
-    internal suspend fun deleteEventById(id: Int) {
+    internal suspend fun deleteEventById(id: Long) {
         repository.deleteEventById(id)
     }
 }
